[Infobulle] La position sur l'écran semble correcte. A tester avec la tablette.

+++ Nomad/src/com/eads/co/nomad/JavaScriptInterface.java
@@ -8,6 +8,6 @@ public class JavaScriptInterface {
 
     
     @android.webkit.JavascriptInterface public void receiveValueFromJs(String str) {
-    	AMMAnnexes.setInfobulle(Integer.parseInt(str));
+    	AMMAnnexes.setInfobulle((int)Double.parseDouble(str));
     }
 }
